Work towards parsing slack input for quotebot.
This changelist implements the parse_slack_input function of the listener module
as well as sets up a unittest for the same. It also includes a refactor of the
command line argument parsing into a function, and the removal of an unnecessary
arg there.

The gitignore file has also been updated to ignore "secrets", so I don't
accidentally upload the API key for the bot or something important like that.