Finish off the basic listener logic.
This commit adds handle_command functionality to the bot, along with unittests.
It also adds a write_response function to take care of writing the messages back
to the channel.

The MagicMock library was added as a dependency in this commit. In the future,
the unittests should be refactored to use the python parameterized library.