Flesh out the quotebot module functionality.
This commit adds logic to the quotebot module to save, retrieve and delete
quotes, as well as tests for the same. It also removes the retrieve specific
quote functionality from the quotebot. If there's a need for this in the future
it can be added, but I think returning random quotes are more fun.