Begin work on the data layer.
After installing and configuring a local postgresql instance, this commit begins
to add some logic to the data layer to perform connections and execute
statements. New files include the schema file, generated from the newly made
database.

The biggest change to the code is the use of a secrets file and configuration
parser to read it. This replaces the cmd line flags, as they would become
unwieldly as the database connection secrets added 3 additional flags.