Update the db schema and work on save functionlity.
This commit makes a change to the database schema, switching the quote_id column
from a non-null integer to a primary key of type "Serial", such that it is
incremented each time a new value is commited to the database.

The data.py file has also been modified to update the save() method with the
functionlity it needs to save quotes to the database, although it is untested as
of yet.