Plumbing through database logic.
This commit adds logic to the quotebot layer to print error message to the slack
channel if any of the data-level operations has failed. It also adds code to the
data.py file to interact with the postgresql database and log/return errors it
might encounter.

The issue outline in bug #2 still applies here though. The next steps will be to
fix that.