Adds insertion functionality to the data layer.
This commit also fixes #2, by formatting the query string properly (specifically
re: the table and column names, which don't need to be the quoted strings that
execute makes them when you pass them in as paramters). The commit also properly
catches the psycopg2 exceptions (rather than just the codes), and updates the
sql query to include the RETURNING directive, which causes the INSERT to compute
and return the quote_id field of the row inserted.