Add Prometheus monitoring to quotebot.
In accordance with issue #12, this commit adds some monitoring statistics to the
quotebot using the prometheus libraries. It adds a counter to record request
counts, which are broken out by each method and the result of calling it (either
success or failure). It also adds a summary metric to record both the count of
times the prefetch_quote_ids function is called, and the amount of time spent in
it, which should be useful for performance profiling.

It also uses the start_http_server function from the prometheus library in the
main() function to boot up an http server for prometheus to scrape.