Add unit tests for the data.py module.
In accordance with issue #6, this change adds some unit tests for data.py. It
mocks the data.connect function by returning instances of a fake class with
methods setup to return known values rather than interacting w/the db. Both
failure and success of each of the three operations are tested.