Fix data.py bug, add postgresql configurations.
This commit fixes a bug in data.py which prevented the QUOTE_LIST from
initializing. It also adds configuration files for the postgresql database,
which need to be updated to connect the quotebot in a docker container to the
database running on the host.

On the host machine, these are symbolically linked from the directory in /etc/
where they normally live to the slackquotebot/postgres_config directory (with
the proper owner and permissions), so that the postgres config will be updated
automatically as any changes to the checked in files are made.